#369 - Renamed file structure, fixed merge conflicts attempt 2
