+++ src/components/projects/project-edit-form/project-edit-container/project-edit-container.tsx
@@ -14,7 +14,7 @@ import ProjectEditDetails from './project-edit-details/project-edit-details';
 import EditModeOptions from './edit-mode-options/edit-mode-options';
 import ProjectEditSummary from './project-edit-summary/project-edit-summary';
 import PageBlock from '../../../shared/page-block/page-block';
-import ChangesList from '../../wbs-details/work-package-container/work-package-container/changes-list/changes-list';
+import ChangesList from '../../wbs-details/work-package-container/work-package-container-view/changes-list/changes-list';
 import { useAllUsers } from '../../../../services/users.hooks';
 import ErrorPage from '../../../shared/error-page/error-page';
 import LoadingIndicator from '../../../shared/loading-indicator/loading-indicator';
