+++ src/components/projects/wbs-details/work-package-container/work-package-container-view/dependencies-list/dependency/dependency.module.css
