+++ src/components/projects/wbs-details/work-package-container/work-package-container-view/work-package-container-view.test.tsx
@@ -5,7 +5,7 @@
 
 import { render, screen, routerWrapperBuilder } from '../../../../../test-support/test-utils';
 import { exampleWorkPackage2 } from '../../../../../test-support/test-data/work-packages.stub';
-import WorkPackageContainer from './work-package-container';
+import WorkPackageContainer from './work-package-container-view';
 import { FormContext } from '../work-package-container';
 
 // Sets up the component under test with the desired values and renders it.
