+++ src/components/projects/wbs-details/work-package-container/work-package-container-view/work-package-details/work-package-details.module.css
